// snl-gms-mock/packages/common-model/src/ts/common/types.ts
export enum StationType {
  SEISMIC_3_COMPONENT = 'SEISMIC_3_COMPONENT',
  SEISMIC_1_COMPONENT = 'SEISMIC_1_COMPONENT',
  SEISMIC_ARRAY = 'SEISMIC_ARRAY',
  HYDROACOUSTIC = 'HYDROACOUSTIC',
  HYDROACOUSTIC_ARRAY = 'HYDROACOUSTIC_ARRAY',
  INFRASOUND = 'INFRASOUND',
  INFRASOUND_ARRAY = 'INFRASOUND_ARRAY',
  WEATHER = 'WEATHER',
  UNKNOWN = 'UNKNOWN'
}

/**
 * Represents Feature Measurement Value for a double type.
 */
export interface DoubleValue {
  readonly value: number;
  readonly standardDeviation?: number;
  readonly units: Units;
}

/**
 * Units used in DoubleValue part of feature prediction and DoubleValue part of calibration
 */
export enum Units {
  DEGREES = 'DEGREES',
  DECIBELS = 'DECIBELS',
  RADIANS = 'RADIANS',
  SECONDS = 'SECONDS',
  HERTZ = 'HERTZ',
  DEGREES_PER_SECOND = 'DEGREES_PER_SECOND',
  SECONDS_PER_DEGREE = 'SECONDS_PER_DEGREE',
  SECONDS_PER_RADIAN = 'SECONDS_PER_RADIAN',
  SECONDS_PER_DEGREE_SQUARED = 'SECONDS_PER_DEGREE_SQUARED',
  SECONDS_PER_KILOMETER_SQUARED = 'SECONDS_PER_KILOMETER_SQUARED',
  SECONDS_PER_KILOMETER = 'SECONDS_PER_KILOMETER',
  SECONDS_PER_KILOMETER_PER_DEGREE = 'SECONDS_PER_KILOMETER_PER_DEGREE',
  ONE_OVER_KM = 'ONE_OVER_KM',
  NANOMETERS = 'NANOMETERS',
  NANOMETERS_PER_SECOND = 'NANOMETERS_PER_SECOND',
  NANOMETERS_PER_COUNT = 'NANOMETERS_PER_COUNT',
  NANOMETERS_SQUARED_PER_SECOND = 'NANOMETERS_SQUARED_PER_SECOND',
  UNITLESS = 'UNITLESS',
  MAGNITUDE = 'MAGNITUDE',
  COUNTS_PER_NANOMETER = 'COUNTS_PER_NANOMETER',
  COUNTS_PER_PASCAL = 'COUNTS_PER_PASCAL',
  PASCALS_PER_COUNT = 'PASCALS_PER_COUNT',
  ONE_OVER_DEGREE = 'ONE_OVER_DEGREE',
  SECONDS_PER_DEGREE_KM = 'SECONDS_PER_DEGREE_KM',
  DEGREES_PER_KM = 'DEGREES_PER_KM'
}